requirements.txt
numpy==1.19.5
pyyaml==6.0
tqdm==4.62.3
scikit-build==0.12.0
torch==1.10.1
torchvision==0.11.2
torchcontrib==0.0.2
pytorch-lightning==1.2.10
scikit-learn==0.24.2
scikit-image==0.17.2
opencv-python==4.5.4.60
Cython
pycocotools==2.0.5
click==8.0.3
kornia==0.6.2